[oidc proxy] Move definitions of functions to source files (#8724)
