+++ ydb/mvp/oidc_proxy/bin/main.cpp
@@ -2,7 +2,7 @@
 
 int main(int argc, char **argv) {
     try {
-        return NMVP::TMVP(argc, argv).Run();
+        return NMVP::NOIDC::TMVP(argc, argv).Run();
     } catch (const yexception& e) {
         Cerr << "Caught exception: " << e.what() << Endl;
         return 1;
