+++ ydb/mvp/oidc_proxy/oidc_session_create_handler.cpp
@@ -3,6 +3,7 @@
 #include "oidc_session_create_yandex.h"
 
 namespace NMVP {
+namespace NOIDC {
 
 TSessionCreateHandler::TSessionCreateHandler(const NActors::TActorId& httpProxyId, const TOpenIdConnectSettings& settings)
     : TBase(&TSessionCreateHandler::StateWork)
@@ -26,4 +27,5 @@ void TSessionCreateHandler::Handle(NHttp::TEvHttpProxy::TEvHttpIncomingRequest::
     ctx.Send(event->Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(response));
 }
 
+}  // NOIDC
 }  // NMVP
