+++ ydb/mvp/oidc_proxy/ya.make
@@ -8,8 +8,15 @@ SRCS(
     mvp.cpp
     oidc_client.cpp
     openid_connect.cpp
+    oidc_settings.cpp
     oidc_protected_page_handler.cpp
+    oidc_protected_page_yandex.cpp
+    oidc_protected_page_nebius.cpp
+    oidc_protected_page.cpp
     oidc_session_create_handler.cpp
+    oidc_session_create_yandex.cpp
+    oidc_session_create_nebius.cpp
+    oidc_session_create.cpp
 )
 
 PEERDIR(
